// deleghebowcl/src/app/model/informativaConsensi.ts
/**
 * DELEGHEBO_API
 * API del back office di Deleghe.
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/*******************************************************************************
* Copyright Regione Piemonte - 2022
* SPDX-License-Identifier: EUPL-1.2
******************************************************************************/

export interface InformativaConsensi {
  pdf_informativa?: string;
  desc_informativa?: string;
  html_informativa?: string
  data_decorrenza?: Date;
  data_scadenza?: Date;
}